Update index.tsx
Seemed to be calling truncate regardless of whether the property was present which caused truncate to call .length against a null, do not have running locally so could not verify in browser if the change was correct, i created a issue for it https://github.com/Mintplex-Labs/vector-admin/issues/65 please verify if my guess was correct. I called an if against that property (data.metadata.text) and just provided place holder text matching the string below if a falsy value was found. Hope i could help.